+++ tests/test_helpers/test_bad_module_attribute_use.py
@@ -502,6 +502,31 @@ class TestBadModuleAttributeUse(dlint.test.base.BaseTest):
 
             assert result == expected
 
+    def test_module_attribute_usage_nested(self):
+        python_node = self.get_ast_node(
+            """
+            import foo
+
+            var = 'echo "TEST"'
+
+            foo.bar(var).baz()
+            """
+        )
+
+        linter = get_bad_module_attribute_use_implementation({'foo': ['bar']})
+        linter.visit(python_node)
+
+        result = linter.get_results()
+        expected = [
+            dlint.linters.base.Flake8Result(
+                lineno=6,
+                col_offset=0,
+                message=linter._error_tmpl
+            )
+        ]
+
+        assert result == expected
+
 
 if __name__ == "__main__":
     unittest.main()